3 - Ollama Chatbot/requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit